Fixed unit test, added integration tests. Decoupled page fetching and added book ammount counting
Noticed I had mixed integration tests and unit tests, so i created a differented file and separated them.
Also page fetcher needed an extension to work for json for the book counting, so added to the class and added the wrapper function for it.
Finally, added function to get to guttendex and look for the ammount of boocks. This will be necesary to automatize the crawler further down the line